python/requirements.txt
PyOpengl
opencv-python
numpy
urchin